+++ processor/src/parallel/tests.rs
@@ -1,5 +1,4 @@
 use alloc::{string::String, sync::Arc};
-use std::string::ToString;
 
 use miden_core::{
     Felt, Kernel, Operation, Program,
@@ -734,5 +733,7 @@ fn external_program() -> Program {
 // See: https://github.com/la10736/rstest/issues/183#issuecomment-1564088329
 #[fixture]
 fn testname() -> String {
-    std::thread::current().name().unwrap().to_string()
+    // Replace `::` with `__` to make snapshot file names Windows-compatible.
+    // Windows does not allow `:` in file names.
+    std::thread::current().name().unwrap().replace("::", "__")
 }